Added a script to find dependencies to install before compile liveplayer Please run ./depends.sh
This script use the file depends-debian
It's possible to make other file for other distributions.
Please read the syntax "help" aviable in the depends-debian file.

After adding a depends-* file, the script must be adjusted.
Comments are existing in the script

To compile this revision:
- cd to the trunk
- run ./depends.sh
- cd src/lp_ladspa_cpp
- qmake && make
- cd ../../
- qmake && make
bin/liveplayer /path/to/song.wav   (or AIFF, or FLAC songs)